pyproject.toml
[project]
name = "feliz_natal"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "kaubarros-24", email = "martinsbarroskaua@gmail.com"},
]
dependencies = ["fastapi>=0.115.6", "uvicorn>=0.34.0", "python-dotenv>=1.0.1", "requests>=2.32.3"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
fast = "uvicorn src.feliz_natal.main:app --port 8080"